+++ examples/resources/nxos_feature_interface_vlan/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_interface_vlan.example "sys/fm/ifvlan"