+++ examples/resources/nxos_feature_interface_vlan/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_interface_vlan" "example" {
+  admin_state = "enabled"
+}