+++ internal/provider/provider.go
@@ -230,6 +230,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_svi_interface":          resourceSVIInterfaceType{},
 		"nxos_svi_interface_vrf":      resourceSVIInterfaceVRFType{},
 		"nxos_feature_ospf":           resourceFeatureOSPFType{},
+		"nxos_feature_interface_vlan": resourceFeatureInterfaceVLANType{},
 	}, nil
 }
 
@@ -245,6 +246,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_svi_interface":          dataSourceSVIInterfaceType{},
 		"nxos_svi_interface_vrf":      dataSourceSVIInterfaceVRFType{},
 		"nxos_feature_ospf":           dataSourceFeatureOSPFType{},
+		"nxos_feature_interface_vlan": dataSourceFeatureInterfaceVLANType{},
 	}, nil
 }