+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add nxos_svi_interface resource and data source
 - Add nxos_svi_interface_vrf resource and data source
+- Add nxos_feature_ospf resource and data source
+- Add nxos_feature_interface_vlan resource and data source
 
 ## 0.2.3